Lights fixes
Soft-start functions were terrible and issued infinity loops because of an uint8_t resetting, so now they work fine.
New signature of the light-switch function: it consumes a pointer to an array with PWMs.
Deleted commented lines for PWM-manipulating without soft-start.
Reworked beamDaySwitch function: now it works as it supposed to.
Added empty function for flashing the high beam, it is in the ToDo-list.

There is a problem with states and switching in a beamSwitch. Also, it's hard to understand it and see an error.
So, I decided to rework this "state machine" to use switch-case operator instead of lots of ifs. That will be in the next commit.