Working features
Now the firmware is totally working (as it seems to me...).
There is no indication, sleep mode, adaptive mode, etc. Just a lights controlling system.
Added "communication.h" that represents some definitions of serial data communication. TX only (for now).
Added "indication.h", not in use. It is supposed to represent the global state on leds, that placed on the pcb.
Changed a bit "pins.h", so there is no collision of one led pin and a digital button's one. Added one more led and a brake sensor pin,
no one is used yet.